adding tests and trying to get circle to work
