+++ .circleci/config.yml
@@ -11,7 +11,7 @@ jobs:
 
       # Tell Circle to store the documentation output in a folder that we can access later
       - store_artifacts:
-          path: _site/
+          path: book_template/_site/
           destination: html
 
   # Deploy the built site to jupyter-book.github.io
@@ -32,7 +32,7 @@ jobs:
       # Grab the the built markdown from the last step
       - run:
           name: Copy over built site files
-          command: rm -rf _build && cp -r /tmp/workspace/_build ./
+          command: cd book_template && rm -rf _build && cp -r /tmp/workspace/_build ./
 
       # Build the site so we have local files
       - build_site:
@@ -41,7 +41,8 @@ jobs:
       # Copying over built website to the live site
       - run:
           name: Copy built site to a folder to store
-          command: cp -r ./_site ../_site
+          command: cp -r ./book_template/_site ../_site
+
       - run:
           name: Change to gh-pages branch and replace w/ new site
           command: |
@@ -57,7 +58,7 @@ jobs:
             git config --global user.name "Chris Holdgraf"
             git add ./.nojekyll
             git add ./* && git commit -m "jekyll book web deploy"
-            git push -f origin gh-pages:gh-pages      
+            git push -f origin gh-pages:gh-pages
 
   run_tests:
     docker:
@@ -74,7 +75,7 @@ jobs:
             - cache-pip
       # Install the packages needed to build our documentation
       # This will depend on your particular package!
-      - run: pip install --user -r build-requirements.txt
+      - run: pip install --user -r requirements.txt
       - run: pip install --user pytest
       # Cache some files for a speedup in subsequent builds
       - save_cache:
@@ -85,20 +86,20 @@ jobs:
       - run:
           name: Tests
           command: |
-            pytest scripts/tests/test_build.py  
+            pytest
 
       # If tests pass, build the site itself
       - run:
           name: Build site markdown
-          command: python scripts/generate_book.py
+          command: cd book_template && python scripts/generate_book.py
 
       # Persist the specified paths (see https://circleci.com/docs/2.0/workflows/#using-workspaces-to-share-data-among-jobs)
       - persist_to_workspace:
           root: .
           paths:
-            - _build/
+            - book_template/_build/
 
-      
+
 
 # Tell CircleCI to use this workflow when it builds the site
 workflows:
@@ -134,11 +135,12 @@ commands:
         default: ""
     steps:
       - checkout
+      - working_directory: book_template
       - restore_cache:
           keys:
             - rails-demo-bundle-v2-{{ checksum "Gemfile.lock" }}
             - rails-demo-bundle-v2-
-      
+
       - run: gem install bundler
       - run: bundle install
 
@@ -151,4 +153,3 @@ commands:
       - run:
           name: Build the website
           command: bundle exec jekyll build << parameters.build_args >>
-
