+++ jupyter_book/tests/site/test_license.md
@@ -1,0 +1,1 @@
+# TEST LICENSE
