+++ requirements.txt
@@ -1,0 +1,2 @@
+pytest
+-r book_template/build-requirements.txt
